+++ lib/application/category_views/all_tools_view.dart
@@ -86,10 +86,10 @@ import 'package:gc_wizard/tools/coords/rhumb_line/widget/rhumbline_waypoint_proj
 import 'package:gc_wizard/tools/coords/triangles/centerofgravity/widget/centerofgravity.dart';
 import 'package:gc_wizard/tools/coords/triangles/circles/excircles/widget/excircles.dart';
 import 'package:gc_wizard/tools/coords/triangles/circles/incircle/widget/incircle.dart';
-import 'package:gc_wizard/tools/coords/triangles/circumcircle/widget/circumcircle.dart';
-import 'package:gc_wizard/tools/coords/triangles/gergonne/widget/gergonne.dart';
-import 'package:gc_wizard/tools/coords/triangles/napoleon/widget/napoleon.dart';
-import 'package:gc_wizard/tools/coords/triangles/orthocenter/widget/orthocenter.dart';
+import 'package:gc_wizard/tools/coords/triangles/circles/circumcircle/widget/circumcircle.dart';
+import 'package:gc_wizard/tools/coords/triangles/special_points/gergonne/widget/gergonne.dart';
+import 'package:gc_wizard/tools/coords/triangles/special_points/napoleon/widget/napoleon.dart';
+import 'package:gc_wizard/tools/coords/triangles/special_points/orthocenter/widget/orthocenter.dart';
 import 'package:gc_wizard/tools/coords/triangles/sidesmidpoint/widget/sidesmidpoint.dart';
 import 'package:gc_wizard/tools/coords/variable_coordinate/widget/variable_coordinate_formulas.dart';
 import 'package:gc_wizard/tools/coords/waypoint_projection/widget/waypoint_projection_geodetic.dart';
