+++ lib/application/category_views/selector_lists/ellipsoid_triangles_selection.dart
@@ -1,15 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:gc_wizard/application/category_views/selector_lists/ellipsoid_triangles_circles_selection.dart';
+import 'package:gc_wizard/application/category_views/selector_lists/ellipsoid_triangles_specialpoints_selection.dart';
 import 'package:gc_wizard/application/registry.dart';
 import 'package:gc_wizard/application/tools/widget/gcw_tool.dart';
 import 'package:gc_wizard/application/tools/widget/gcw_toollist.dart';
 import 'package:gc_wizard/common_widgets/gcw_selection.dart';
 import 'package:gc_wizard/tools/coords/triangles/centerofgravity/widget/centerofgravity.dart';
-import 'package:gc_wizard/tools/coords/triangles/circles/excircles/widget/excircles.dart';
-import 'package:gc_wizard/tools/coords/triangles/circles/incircle/widget/incircle.dart';
-import 'package:gc_wizard/tools/coords/triangles/circumcircle/widget/circumcircle.dart';
-import 'package:gc_wizard/tools/coords/triangles/gergonne/widget/gergonne.dart';
-import 'package:gc_wizard/tools/coords/triangles/napoleon/widget/napoleon.dart';
-import 'package:gc_wizard/tools/coords/triangles/orthocenter/widget/orthocenter.dart';
 import 'package:gc_wizard/tools/coords/triangles/sidesmidpoint/widget/sidesmidpoint.dart';
 import 'package:gc_wizard/utils/ui_dependent_utils/common_widget_utils.dart';
 
@@ -20,14 +16,10 @@ class EllipsoidTrianglePointsSelection extends GCWSelection {
   Widget build(BuildContext context) {
     final List<GCWTool> _toolList = registeredTools.where((element) {
       return [
-        className(const TriangleNapoleonPoints()),
+        className(const EllipsoidTrianglePointsCirclesSelection()),
         className(const TriangleSideMidPoints()),
         className(const TriangleCenterOfGravity()),
-        className(const TriangleIncircle()),
-        className(const TriangleCircumCircle()),
-        className(const TriangleOrthocenter()),
-        className(const TriangleExcircles()),
-        className(const TriangleGergonnePoint()),
+        className(const EllipsoidTrianglePointsSpecialPointsSelection()),
       ].contains(className(element.tool));
     }).toList();
