+++ lib/application/registry.dart
@@ -30,7 +30,9 @@ import 'package:gc_wizard/application/category_views/selector_lists/dna_selectio
 import 'package:gc_wizard/application/category_views/selector_lists/e_selection.dart';
 import 'package:gc_wizard/application/category_views/selector_lists/easter_selection.dart';
 import 'package:gc_wizard/application/category_views/selector_lists/elements_of_geocaching_selection.dart';
+import 'package:gc_wizard/application/category_views/selector_lists/ellipsoid_triangles_circles_selection.dart';
 import 'package:gc_wizard/application/category_views/selector_lists/ellipsoid_triangles_selection.dart';
+import 'package:gc_wizard/application/category_views/selector_lists/ellipsoid_triangles_specialpoints_selection.dart';
 import 'package:gc_wizard/application/category_views/selector_lists/esoteric_programminglanguages_selection.dart';
 import 'package:gc_wizard/application/category_views/selector_lists/games_selection.dart';
 import 'package:gc_wizard/application/category_views/selector_lists/general_codebreakers_selection.dart';
@@ -154,10 +156,10 @@ import 'package:gc_wizard/tools/coords/segment_line/widget/segment_line.dart';
 import 'package:gc_wizard/tools/coords/triangles/centerofgravity/widget/centerofgravity.dart';
 import 'package:gc_wizard/tools/coords/triangles/circles/excircles/widget/excircles.dart';
 import 'package:gc_wizard/tools/coords/triangles/circles/incircle/widget/incircle.dart';
-import 'package:gc_wizard/tools/coords/triangles/circumcircle/widget/circumcircle.dart';
-import 'package:gc_wizard/tools/coords/triangles/gergonne/widget/gergonne.dart';
-import 'package:gc_wizard/tools/coords/triangles/napoleon/widget/napoleon.dart';
-import 'package:gc_wizard/tools/coords/triangles/orthocenter/widget/orthocenter.dart';
+import 'package:gc_wizard/tools/coords/triangles/circles/circumcircle/widget/circumcircle.dart';
+import 'package:gc_wizard/tools/coords/triangles/special_points/gergonne/widget/gergonne.dart';
+import 'package:gc_wizard/tools/coords/triangles/special_points/napoleon/widget/napoleon.dart';
+import 'package:gc_wizard/tools/coords/triangles/special_points/orthocenter/widget/orthocenter.dart';
 import 'package:gc_wizard/tools/coords/triangles/sidesmidpoint/widget/sidesmidpoint.dart';
 import 'package:gc_wizard/tools/coords/variable_coordinate/widget/variable_coordinate_formulas.dart';
 import 'package:gc_wizard/tools/coords/waypoint_projection/widget/waypoint_projection_geodetic.dart';
@@ -4024,6 +4026,18 @@ void initializeRegistry(BuildContext context) {
       ]),
 
     // EllipsoidTrianglePointsSelection ***************************************************
+    GCWTool(
+        tool: const EllipsoidTrianglePointsCirclesSelection(),
+        id: 'triangle_ellipsoid_circles_selection',
+        searchKeys: const [
+          'triangle',
+        ]),
+    GCWTool(
+        tool: const EllipsoidTrianglePointsSpecialPointsSelection(),
+        id: 'triangle_ellipsoid_specialpoints_selection',
+        searchKeys: const [
+          'triangle',
+        ]),
     GCWTool(
         tool: const TriangleNapoleonPoints(),
         id: 'triangle_point_napoleon',
