+++ lib/application/theme/fixed_colors.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 const COLOR_MAP_POLYLINE = Colors.red;
+const COLOR_MAP_POLYLINE_TRIANGLE = Colors.black;
 const COLOR_MAP_CIRCLE = Colors.blue;
 const COLOR_MAP_POINT = Colors.black;
 const COLOR_MAP_CALCULATEDPOINT = Color.fromRGBO(6, 140, 0, 1.0);
