+++ test/tools/coords/triangles/circles/excircles/logic/excircles_test.dart
@@ -15,8 +15,20 @@ void main() async {
     for (int i = 0; i < validEllipsoidTrianglesForTests.length; i++) {
       var triangleTest = validEllipsoidTrianglesForTests[i];
 
-      var triangle = ELlipsoidTriangle(triangleTest['inputA'] as LatLng, triangleTest['inputB'] as LatLng, triangleTest['inputC'] as LatLng, Ellipsoid.WGS84);
+      var triangle = EllipsoidTriangle(triangleTest['inputA'] as LatLng, triangleTest['inputB'] as LatLng, triangleTest['inputC'] as LatLng, Ellipsoid.WGS84);
       test('input: $triangle, cw: ${triangle.isClockwise}', () {
+        if (
+             (triangle.bearingAB == 0 && triangle.bearingBA == 0)
+          || (triangle.bearingAB == 180 && triangle.bearingBA == 180)
+          || (triangle.bearingBC == 0 && triangle.bearingCB == 0)
+          || (triangle.bearingBC == 180 && triangle.bearingCB == 180)
+          || (triangle.bearingAC == 0 && triangle.bearingCA == 0)
+          || (triangle.bearingAC == 180 && triangle.bearingCA == 180)
+        ) {
+          // Extreme edge case where at least one side goes over pole, accepted risk (SMan, 04/2026)
+          return;
+        }
+
         var _actual = calculateEllipsoidTriangleExcircles(triangle, Ellipsoid.WGS84);
         if (_actual == null) {
           expect(_actual == null, !triangle.isValid || triangleIsMeridianCircle(triangle, Ellipsoid.WGS84));
@@ -44,7 +56,7 @@ void main() async {
           expect(((dists[2] + dists[0]) / 2 - excircle.circle.radius).abs() < 1e-6, true);
 
           // Not an Incircle
-          expect(utils.equalsLatLng(excircle.circle.center, incircle!.center, tolerance: 1e-6), false);
+          expect(utils.equalsLatLng(excircle.circle.center, incircle!.circle.center, tolerance: 1e-6), false);
         }
       });
     }
