--- test/tools/coords/triangles/gergonne/logic/gergonne_test.dart
@@ -1,32 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-import 'package:gc_wizard/tools/coords/_common/logic/ellipsoid.dart';
-import 'package:gc_wizard/tools/coords/triangles/_common/logic/ellipsoid_triangle.dart';
-import 'package:gc_wizard/tools/coords/triangles/gergonne/logic/gergonne.dart';
-import 'package:latlong2/latlong.dart';
-
-import '../../../../science_and_technology/euclidic_triangle/logic/triangle_test_utils.dart';
-
-void main() async {
-  group("triangle.calculateEllipsoidTriangleGergonnePoint:", () {
-    List<Map<String, Object?>> _inputsToExpected = [
-      {'inputA': LatLng(0, 0), 'inputB': LatLng(0, 0), 'inputC': LatLng(0, 0),
-        'expectedOutput': LatLng(double.nan, double.nan)},
-      {'inputA': LatLng(1, 1), 'inputB': LatLng(1, 1), 'inputC': LatLng(1, 1),
-        'expectedOutput': LatLng(double.nan, double.nan)},
-      {'inputA': LatLng(1, 1), 'inputB': LatLng(2, 2), 'inputC': LatLng(3, 3),
-        'expectedOutput': LatLng(double.nan, double.nan)},
-      {'inputA': LatLng(0, 0), 'inputB': LatLng(0, 3), 'inputC': LatLng(4, 0),
-        'expectedOutput': LatLng(0.7317379990783085, 0.8157721928407682)},
-      {'inputA': LatLng(40, 9), 'inputB': LatLng(42, 9), 'inputC': LatLng(38, 8),
-        'expectedOutput': LatLng(40.0025888794003, 8.98268272877226)},
-    ];
-
-    for (var elem in _inputsToExpected) {
-      test('input: ${elem['inputA']} ${elem['inputB']} ${elem['inputC']}', () {
-        var triangle = ELlipsoidTriangle(elem['inputA'] as LatLng, elem['inputB'] as LatLng, elem['inputC'] as LatLng, Ellipsoid.WGS84);
-        var _actual = calculateEllipsoidTriangleGergonnePoint(triangle, Ellipsoid.WGS84);
-        latLngTest(_actual!, elem['expectedOutput'] as LatLng);
-      });
-    }
-  });
-}
